+++ src/pml_vqvae/models/baseline/vae.py
@@ -12,7 +12,7 @@ from torch import nn
 
 from pml_vqvae.visuals import show
 
-from pml_vqvae.models.baseline.pml_model_interface import PML_model
+from pml_vqvae.models.pml_model_interface import PML_model
 from pml_vqvae.nnutils import ResidualBlock
